poc/requirements.txt
cryptography
dnspython
